+++ dj_twilio_sms/serializers.py
@@ -6,21 +6,23 @@ from .models import IncomingSMS, OutgoingSMS
 
 
 class SMSRequestSerializer(serializers.ModelSerializer):
-    SmsSid = serializers.CharField(max_length=34, required=True, source="sms_sid")
+    MessageSid = serializers.CharField(max_length=34, required=True, source="sms_sid")
     AccountSid = serializers.CharField(max_length=34, required=True, source="account_sid")
     From = serializers.CharField(max_length=30, required=True, source="from_number")
-    FromCity = serializers.CharField(max_length=30, required=False, source="from_city")
-    FromState = serializers.CharField(max_length=30, required=False, source="from_state")
-    FromZip = serializers.CharField(max_length=30, required=False, source="from_zip")
-    FromCountry = serializers.CharField(max_length=120, required=False, source="from_country")
     To = serializers.CharField(max_length=30, required=True, source="to_number")
     Body = serializers.CharField(max_length=160, required=True, source="body")
+    FromCity = serializers.CharField(max_length=100, required=False, source="from_city")
+    FromState = serializers.CharField(max_length=100, required=False, source="from_state")
+    FromZip = serializers.CharField(max_length=30, required=False, source="from_zip")
+    FromCountry = serializers.CharField(max_length=120, required=False, source="from_country")
+
+    # @TODO Add MessagingServiceSid and NumMedia
 
     class Meta:
         model = IncomingSMS
         fields = [
-            "SmsSid", "AccountSid", "From", "FromCity", "FromState", "FromZip",
-            "FromCountry", "To", "Body"
+            "MessageSid", "AccountSid", "From", "To", "Body",
+            "FromCity", "FromState", "FromZip", "FromCountry"
         ]
 
 
@@ -31,5 +33,5 @@ class SMSStatusSerializer(serializers.ModelSerializer):
     class Meta:
         model = OutgoingSMS
         fields = [
-            "SmsSid", "SmsStatus",
+            "SmsSid", "SmsStatus"
         ]